requirements.txt
torch==1.9.0
torchvision==0.10.0
Pillow==8.2.0
Flask==2.0.1
tensorflow==2.5.0
